REFACTOR: STEP2 구조 및 코드 수정
